+++ src/main/java/subway/service/dto/request/SearchAllSectionLineRequest.java
@@ -1,4 +1,4 @@
-package subway.service.dto;
+package subway.service.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
