+++ src/main/java/subway/service/dto/response/SectionInLineResponse.java
@@ -1,4 +1,4 @@
-package subway.service.dto;
+package subway.service.dto.response;
 
 public class SectionInLineResponse {
